+++ components-experimental/OldSlider.vue
