+++ components-experimental/Slider.vue
